wip: changed according to discussion with team, formatted the excel template

+++ idas.chox.data/src/main/java/idas/chox/data/ExcelInvoice.java
@@ -176,9 +176,9 @@ public class ExcelInvoice {
         claimsHandlingInvoiceAmount = (BigDecimal) data.get("claimshandlinginvoiceamount");
         claimInvoiceNo = (String) data.get("claiminvoiceno");
         BigDecimal temphireRateOriginalDays = (BigDecimal) data.get("orignaldays");
-        Integer tempActualDays = (Integer) data.get("days");
+        BigDecimal tempActualDays = (BigDecimal) data.get("days");
         if (temphireRateOriginalDays == null) {
-            hireRateOriginalDays = (Integer) tempActualDays;
+            hireRateOriginalDays = tempActualDays.intValue();
         } else {
             hireRateOriginalDays = temphireRateOriginalDays.intValue();
         }
